+++ src/main/java/org/dmg/storyteller/ContextLike.scala
@@ -1,4 +1,4 @@
-package org.dmg.storyteller.fact
+package org.dmg.storyteller
 
 trait ContextLike {
   def has(expression: String): Boolean
