+++ src/main/java/org/dmg/storyteller/GlobalContext.scala
@@ -1,4 +1,4 @@
-package org.dmg.storyteller.fact
+package org.dmg.storyteller
 
 import scala.collection.mutable.ListBuffer
 import scala.language.{implicitConversions, postfixOps}
